HIRISE_api/requirements.txt
beautifulsoup4
livelossplot
humanize
pandas
numpy
torch
torchvision
opencv-python
math
PIL
rasterfairy
seaborn
hdbscan
umap-learn
pytest
requests
tqdm
twill
setuptools
sklearn
mlxtend
pytest